Prep for homelab deploy; add tracing to initial world generation
