+++ docker/.local.env
@@ -1,0 +1,3 @@
+MYSQL_USER=root
+MYSQL_PASS=dev
+MYSQL_HOST=mysql
