+++ docker/homelab-deploy-compose.yaml
@@ -7,9 +7,14 @@ services:
     environment:
       - OTEL_ENDPOINT=http://otel-collector:4317
       - RUST_BACKTRACE=1
+      - MYSQL_DATABASE=dax_director
     ports:
       - "6081:6081"
+    volumes:
+      - director-migrations:/migrations
     depends_on:
+      run-svc-migrations:
+        condition: service_completed_successfully
       valkey:
         condition: service_started
   webapp:
@@ -25,6 +30,8 @@ services:
     ports:
       - "6080:6080"
     depends_on:
+      run-svc-migrations:
+        condition: service_completed_successfully
       valkey:
         condition: service_started
   auth:
@@ -38,6 +45,8 @@ services:
     ports:
       - "6082:6082"
     depends_on:
+      run-svc-migrations:
+        condition: service_completed_successfully
       valkey:
         condition: service_started
   get-openobserve-secret:
@@ -116,6 +125,49 @@ services:
       - "http://webapp:6080"
       - "--token"
       - "${CLOUDFLARED_TOKEN}"
+  mysql:
+    container_name: "mysql"
+    image: "mysql:9"
+    environment:
+      - MYSQL_ROOT_PASSWORD=${MYSQL_PASS}
+    healthcheck:
+      test: mysqladmin ping -h ${MYSQL_HOST} -u${MYSQL_USER} -p${MYSQL_PASS} || exit 1
+      interval: 10s
+      retries: 5
+      timeout: 25s
+  run-svc-migrations:
+    container_name: "run-svc-migrations"
+    image: "rust:1.88-bookworm"
+    volumes:
+      - director-migrations:/migrations/director:ro
+    command:
+      - "/bin/bash"
+      - "-c"
+      - |
+        cargo install sqlx-cli
+
+        echo "find results = $(find /migrations -type d -not -name migrations)"
+        for migration_dir in $(find /migrations -type d -not -name migrations);
+        do
+          echo "Running following migrations in $${migration_dir}..."
+          database_name=$(echo $${migration_dir} | sed -E "s/\/migrations\/(.*)/\1/g")
+          echo "Creating database $${database_name}..."
+          echo $(cat $${migration_dir}/*.sql)
+          sqlx database create \
+            --database-url=mysql://${MYSQL_USER}:${MYSQL_PASS}@${MYSQL_HOST}/dax_$${database_name} \
+          || exit 1
+
+          echo "Running migrations..."
+          sqlx migrate run \
+            --database-url=mysql://${MYSQL_USER}:${MYSQL_PASS}@${MYSQL_HOST}/dax_$${database_name} \
+            --source $${migration_dir} \
+          || exit 1
+        done
+        exit 0
+    depends_on:
+      mysql:
+        condition: service_healthy
 volumes:
+  director-migrations:
   redisinsight:
   otelcollector:
